pyproject.toml
[tool.poetry]
name = "hls-transcoder"
version = "0.1.0"
description = "Convert video files to HLS format with hardware acceleration and parallel processing"
authors = ["Souvik Karmakar <souvikk431@gmail.com>"]
readme = "README.md"
license = "MIT"
packages = [{ include = "hls_transcoder" }]

[tool.poetry.dependencies]
python = "^3.10"
ffmpeg-python = "^0.2.0"
typer = { extras = ["all"], version = "0.20.0" }
rich = "^14.2.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"
pytest-cov = "^7.0.0"
black = "^25.9.0"
mypy = "^1.7.0"
ruff = "^0.14.3"
typer = "^0.20.0"

[tool.poetry.scripts]
hls-transcoder = "hls_transcoder.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
